requirements.txt
numpy==1.23.5
pandas==1.5.3
python-dotenv==1.0.0

matplotlib==3.7.1
plotly==5.15.0

scikit-learn==1.2.2
scipy==1.10.1
xgboost==1.7.5
lightgbm==3.3.5

mlflow==2.3.1
boto3==1.28.59
pyarrow<12
protobuf==3.20.3

jupyterlab==3.6.5
ipykernel==6.25.2

tqdm==4.66.1
